chore(discover-quick-context): Fixed type errors.
